Plushie Update! (#654)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

By multiple requests i bring my PR from Frontier The [PLUSHIE
UPDATE!](https://github.com/new-frontiers-14/frontier-station-14/pull/886)

This PR add a prize vendor machine where you can trade prize ticket for
a toy/plushie, to get ticket you win those in arcade machines.

This PR will give a chance for everyone to get the toys they need in a
special machine and will make plushies in general more random and fun,
in term of random spawns.

WARNING THIS PR DOES NOT MAP THE PRIZE VENDOR AND WILL MOSTLY NEED TO BE
MAPPED.

---

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Prize Vendor
- [x] Prize Tickets
- [x] New Toys/Plushies
- [x] Prizes

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/643e6c65-bba0-4eef-8e8e-d3510922e51e)

</p>
</details>

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: A ton of new plushies has arrived
- add: Prize counter are here, get your prizes by winning games and
getting prize tickets! and find prize balls!!